requirements.txt
apache-airflow==2.10.3
pyspark==3.4.1
delta-spark==2.4.0
mlflow==2.14.0
pymongo==4.10.1
pandas==2.2.3
numpy==1.26.4
optuna==4.6.0
optuna-integration==4.6.0